Refactor: Improve code organization and naming conventions
