+++ data/local/src/main/kotlin/com/berlin/local/datasource/SearchLocalDataSourceImpl.kt
@@ -2,8 +2,8 @@ package com.berlin.local.datasource
 
 import com.berlin.local.dao.SearchDao
 import com.berlin.repository.datasource.local.datasource.SearchLocalDataSource
-import com.berlin.repository.datasource.local.dto.QueryType
 import com.berlin.repository.datasource.local.dto.RecentSearchHistoryEntity
+import com.berlin.repository.util.QueryType
 import javax.inject.Inject
 
 class SearchLocalDataSourceImpl  @Inject constructor (
